requirements.txt
# Simulation
simpy==4.1.1

# Data processing
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# Configuration
pyyaml==6.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
Pillow==10.2.0

# Animation
imageio==2.34.0
imageio-ffmpeg==0.4.9

# Web application
streamlit==1.31.0

# Utilities
tqdm==4.66.1

# OCR (optional - for timetable image processing)
pytesseract==0.3.10

# Layout editor (optional - for clicking on image to set coordinates)
streamlit-image-coordinates==0.2.2

